server/requirements.txt
fastapi
uvicorn
pydantic
httpx
groq
python-dotenv